synchronous printer is getting so close!

+++ PropWare/PropWare.h
@@ -50,7 +50,7 @@ namespace PropWare {
 #define MILLISECOND         ((uint32_t) (CLKFREQ / 1000))
 #define MICROSECOND         ((uint32_t) (MILLISECOND / 1000))
 
-typedef int8_t ErrorCode;
+typedef int16_t ErrorCode;
 
 typedef enum {
     BIT_0  = 0x1,
